Full app usage statistics module; split app and profile model classes into data and domain for simplicity; refactor

+++ app/src/main/java/com/darkzodiak/kontrol/statistics/presentation/components/AppUsage.kt
@@ -49,7 +49,7 @@ fun AppUsage(
 }
 
 @Composable
-fun formatPercent(percent: Long): String {
+fun formatPercent(percent: Int): String {
     return if (percent > 0) "$percent %"
         else "< 1 %"
 }
